+++ src/main/java/com/seoulmilk/seoulmilkServer/domain/ntsTax/domain/ARAP.java
@@ -1,0 +1,5 @@
+package com.seoulmilk.seoulmilkServer.domain.ntsTax.domain;
+
+public enum ARAP {
+    AR, AP
+}
