--- src/main/java/com/seoulmilk/seoulmilkServer/domain/ntsTax/domain/enums/ARAP.java
@@ -1,5 +1,0 @@
-package com.seoulmilk.seoulmilkServer.domain.ntsTax.domain.enums;
-
-public enum ARAP {
-    AR, AP
-}
