--- src/main/java/com/seoulmilk/seoulmilkServer/domain/ntsTax/domain/enums/TypeCode.java
@@ -1,4 +1,0 @@
-package com.seoulmilk.seoulmilkServer.domain.ntsTax.domain.enums;
-
-public enum TypeCode {
-}
